requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiogram-unittest==0.0.1
aiogram==2.24
aiohttp==3.8.3; python_version >= '3.6'
aioschedule==0.5.2
aiosignal==1.3.1; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
attrs==22.2.0; python_version >= '3.6'
babel==2.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
certifi==2022.12.7; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
frozenlist==1.3.3; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
magic-filter==1.0.9; python_version >= '3.7' and python_version < '4.0'
multidict==6.0.4; python_version >= '3.7'
python-dotenv==0.21.1
pytz==2022.7.1
yarl==1.8.2; python_version >= '3.7'
